% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codependent-package.R
\docType{package}
\name{codependent-package}
\alias{codependent-package}
\alias{codependent}
\title{Rarefaction and Extrapolation on Bipartite Species Association Networks}
\description{
The codependent package iteratively subsamples bipartite networks, to estimate the scaling of host and affiliate richness in ecological networks.
}
\author{
Colin Carlson (\email{ccarlson@sesync.org})
}
