% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coextinct.R
\name{coextinct}
\alias{coextinct}
\title{Binary rarefaction (curve-fitting)}
\usage{
coextinct(assoc.df, iter = 100, plots = FALSE, subSample = NULL)
}
\arguments{
\item{assoc.df}{The raw dataset of associations (pairs of names in two columns, hosts and affiliates)}

\item{iter}{Number of iterations required for each subsample}

\item{plots}{Do you want plots? Maybe you do. The world is your oyster.}

\item{subSample}{Fraction of host species to subsample (default=NULL, use all data)}
}
\description{
This function runs the subsampling algorithm on the association dataset, and returns a fitted model describing the power law relationship
}
