% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvedf50.R, R/utils.R
\name{curve.50}
\alias{curve.50}
\alias{curve.50}
\title{Subsampling of 50\% o network}
\usage{
curve.50(assoc.df, iter)

curve.50(assoc.df, iter)
}
\arguments{
\item{assoc.df}{The raw dataset of associations (pairs of names in two columns, hosts and affiliates)}

\item{iter}{Number of iterations required for each subsample}

\item{assoc.df}{A data frame of host-affiliate intractions, host names in the first column and affiliate names in the second}

\item{iter}{}

\item{hostList}{currently undefined, but assumed to be global?}
}
\description{
This function iteratively samples half the network and generates a data frame with sub-sample host and affilaite richness tallied together


}
