% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigroup.R
\name{multigroup}
\alias{multigroup}
\title{Multigroup estimates}
\usage{
multigroup(fake, orders = c("A", "B", "C", "D", "E", "F"), est.num.fake)
}
\arguments{
\item{fake}{A data frame of the group membership of all the sample species}

\item{orders}{You gotta just quickly name all the groups}

\item{est.num.fake}{The numbers you extrapolated for each group}
}
\description{
This function runs the new inclusion-exclusion algorithm correction designed by Colin Carlson and Laura Alexander and implemented in Carlson et al. (in prep)
}
