% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zstar.R
\name{zstar}
\alias{zstar}
\title{Asymptotic scaling estimator}
\usage{
zstar(assoc.df, nsamp = 1, npts = 100, plot = TRUE)
}
\arguments{
\item{hell}{ass}
}
\description{
Estimating z values asymptotically
}
\details{

}
