% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvedf50.R
\name{curve.50}
\alias{curve.50}
\title{Subsampling of 50\% o network}
\usage{
curve.50(assoc.df, iter)
}
\arguments{
\item{assoc.df}{The raw dataset of associations (pairs of names in two columns, hosts and affiliates)}

\item{iter}{Number of iterations required for each subsample}
}
\description{
This function iteratively samples half the network and generates a data frame with sub-sample host and affilaite richness tallied together
}
